+ Extensions context, separate from Connect
