+++ src/Providers.tsx
@@ -1,36 +1,37 @@
 // Copyright 2022 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import { AccountProvider } from 'contexts/Account';
+import { APIProvider, useApi } from 'contexts/Api';
+import { BalancesProvider } from 'contexts/Balances';
+import { ConnectProvider } from 'contexts/Connect';
+import { ExtensionsProvider } from 'contexts/Extensions';
+import { ExtrinsicsProvider } from 'contexts/Extrinsics';
+import { FiltersProvider } from 'contexts/Filters';
+import { HelpProvider } from 'contexts/Help';
+import { MenuProvider } from 'contexts/Menu';
+import { ModalProvider } from 'contexts/Modal';
+import { NetworkMetricsProvider } from 'contexts/Network';
+import { NotificationsProvider } from 'contexts/Notifications';
+import { OverlayProvider } from 'contexts/Overlay';
+import { ActivePoolsProvider } from 'contexts/Pools/ActivePools';
+import { BondedPoolsProvider } from 'contexts/Pools/BondedPools';
+import { PoolMembersProvider } from 'contexts/Pools/PoolMembers';
+import { PoolMembershipsProvider } from 'contexts/Pools/PoolMemberships';
+import { PoolsConfigProvider } from 'contexts/Pools/PoolsConfig';
+import { SessionEraProvider } from 'contexts/SessionEra';
+import { StakingProvider } from 'contexts/Staking';
+import { SubscanProvider } from 'contexts/Subscan';
+import { useTheme } from 'contexts/Themes';
 import { TooltipProvider } from 'contexts/Tooltip';
 import { TransferOptionsProvider } from 'contexts/TransferOptions';
 import { TxFeesProvider } from 'contexts/TxFees';
+import { UIProvider } from 'contexts/UI';
+import { ValidatorsProvider } from 'contexts/Validators';
 import { withProviders } from 'library/Hooks';
 import Router from 'Router';
 import { ThemeProvider } from 'styled-components';
 import { EntryWrapper as Wrapper } from 'Wrappers';
-import { AccountProvider } from './contexts/Account';
-import { APIProvider, useApi } from './contexts/Api';
-import { BalancesProvider } from './contexts/Balances';
-import { ConnectProvider } from './contexts/Connect';
-import { ExtrinsicsProvider } from './contexts/Extrinsics';
-import { FiltersProvider } from './contexts/Filters';
-import { HelpProvider } from './contexts/Help';
-import { MenuProvider } from './contexts/Menu';
-import { ModalProvider } from './contexts/Modal';
-import { NetworkMetricsProvider } from './contexts/Network';
-import { NotificationsProvider } from './contexts/Notifications';
-import { OverlayProvider } from './contexts/Overlay';
-import { ActivePoolsProvider } from './contexts/Pools/ActivePools';
-import { BondedPoolsProvider } from './contexts/Pools/BondedPools';
-import { PoolMembersProvider } from './contexts/Pools/PoolMembers';
-import { PoolMembershipsProvider } from './contexts/Pools/PoolMemberships';
-import { PoolsConfigProvider } from './contexts/Pools/PoolsConfig';
-import { SessionEraProvider } from './contexts/SessionEra';
-import { StakingProvider } from './contexts/Staking';
-import { SubscanProvider } from './contexts/Subscan';
-import { useTheme } from './contexts/Themes';
-import { UIProvider } from './contexts/UI';
-import { ValidatorsProvider } from './contexts/Validators';
 
 // `polkadot-dashboard-ui` theme classes are inserted here.
 export const WrappedRouter = () => {
@@ -61,6 +62,7 @@ export const ThemedRouter = () => {
 export const Providers = withProviders(
   FiltersProvider,
   APIProvider,
+  ExtensionsProvider,
   ConnectProvider,
   HelpProvider,
   NetworkMetricsProvider,
